Move Box components to common and Login to pages
