+++ assets/js/common/LoadingBox/index.js
@@ -1,0 +1,3 @@
+import LoadingBox from './LoadingBox';
+
+export default LoadingBox;
