+++ assets/js/common/NotificationBox/NotificationBox.jsx
