+++ assets/js/common/NotificationBox/index.js
@@ -1,0 +1,3 @@
+import NotificationBox from './NotificationBox';
+
+export default NotificationBox;
