+++ assets/js/pages/HostDetails/HostSettingsPage.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import LoadingBox from '@components/LoadingBox';
+import LoadingBox from '@common/LoadingBox';
 
 import { TARGET_HOST } from '@lib/model';
